added generic default parser
